Review changes to move test-specific constants to test are
