+++ src/test/java/uk/gov/hmcts/reform/bulkscan/utils/Constants.java
@@ -10,6 +10,19 @@ public final class Constants {
     public static final String APPLICANT_1_FIRST_NAME = "applicant1_firstName";
     public static final String APPLICANT_1_LAST_NAME = "applicant1_lastName";
 
+    public static final String EXEMPTION_TO_ATTEND_MIAM_GROUP_FIELD = "exemption_to_attend_MIAM";
+    public static final String NOMIAM_CHILDPROTECTIONCONCERNS_FIELD = "NoMIAM_childProtectionConcerns";
+    public static final String NOMIAM_DOMESTICVIOLENCE_FIELD = "NoMIAM_domesticViolence";
+    public static final String NOMIAM_DVE_ARRESTEDFORSIMILAROFFENCE_FIELD = "NoMIAM_DVE_arrestedForSimilarOffence";
+    public static final String NOMIAM_URGENCY_FIELD = "NoMIAM_Urgency";
+    public static final String NOMIAM_PREVIOUSATTENDANCE_FIELD = "NoMIAM_PreviousAttendance";
+    public static final String NOMIAM_OTHERREASONS_FIELD = "NoMIAM_otherReasons";
+
+    public static final String TICK_BOX_FALSE = "false";
+    public static final String TICK_BOX_NO = "No";
+    public static final String TICK_BOX_TRUE = "true";
+    public static final String TICK_BOX_YES = "Yes";
+
     private Constants() {
     }
 }
